+++ Sources/StreamChat/APIClient/CDNClient/CDNClient.swift
@@ -51,6 +51,15 @@ public protocol CDNClient {
         progress: ((Double) -> Void)?,
         completion: @escaping (Result<UploadedFile, Error>) -> Void
     )
+
+    /// Deletes the attachment from the CDN, given the remote URL.
+    /// - Parameters:
+    ///   - remoteUrl: The remote url of the attachment.
+    ///   - completion: Returns an error in case the delete operation fails.
+    func deleteAttachment(
+        remoteUrl: URL,
+        completion: @escaping (Error?) -> Void
+    )
 }
 
 public extension CDNClient {
@@ -147,7 +156,39 @@ class StreamCDNClient: CDNClient {
             completion: completion
         )
     }
-    
+
+    func deleteAttachment(
+        remoteUrl: URL,
+        completion: @escaping (Error?) -> Void
+    ) {
+        let isImage = AttachmentFileType(ext: remoteUrl.pathExtension).isImage
+        let endpoint = Endpoint<EmptyResponse>
+            .deleteAttachment(
+                url: remoteUrl,
+                type: isImage ? .image : .file
+            )
+
+        encoder.encodeRequest(for: endpoint) { [weak self] (requestResult) in
+            var urlRequest: URLRequest
+
+            do {
+                urlRequest = try requestResult.get()
+            } catch {
+                log.error(error, subsystems: .httpRequests)
+                completion(error)
+                return
+            }
+
+            guard let self = self else {
+                return
+            }
+
+            self.session.dataTask(with: urlRequest, completionHandler: { _, _, error in
+                completion(error)
+            }).resume()
+        }
+    }
+
     private func uploadAttachment<ResponsePayload>(
         endpoint: Endpoint<ResponsePayload>,
         fileData: Data,