+++ Sources/StreamChat/WebSocketClient/WebSocketClient.swift
@@ -42,7 +42,7 @@ class WebSocketClient {
     ///
     /// Changing this value doesn't automatically update the existing connection. You need to manually call `disconnect`
     /// and `connect` to make a new connection to the updated endpoint.
-    var connectEndpoint: Endpoint<EmptyResponse>?
+    @Atomic var connectEndpoint: Endpoint<EmptyResponse>?
 
     /// The decoder used to decode incoming events
     private let eventDecoder: AnyEventDecoder
@@ -121,7 +121,9 @@ class WebSocketClient {
         }
 
         do {
-            engine = try createEngineIfNeeded(for: endpoint)
+            try engineQueue.sync {
+                self.engine = try createEngineIfNeeded(for: endpoint)
+            }
         } catch {
             return
         }