+++ TestTools/StreamChatTestTools/Mocks/StreamChat/Workers/ChannelUpdater_Mock.swift
@@ -14,10 +14,12 @@ final class ChannelUpdater_Mock: ChannelUpdater {
 
     @Atomic var updateChannel_payload: ChannelEditDetailPayload?
     @Atomic var updateChannel_completion: ((Error?) -> Void)?
+    @Atomic var updateChannel_completion_result: Result<Void, Error>?
 
     @Atomic var partialChannelUpdate_updates: ChannelEditDetailPayload?
     @Atomic var partialChannelUpdate_unsetProperties: [String]?
     @Atomic var partialChannelUpdate_completion: ((Error?) -> Void)?
+    @Atomic var partialChannelUpdate_completion_result: Result<Void, Error>?
 
     @Atomic var muteChannel_cid: ChannelId?
     @Atomic var muteChannel_expiration: Int?
@@ -102,7 +104,7 @@ final class ChannelUpdater_Mock: ChannelUpdater {
 
     @Atomic var markUnread_cid: ChannelId?
     @Atomic var markUnread_userId: UserId?
-    @Atomic var markUnread_messageId: MessageId?
+    @Atomic var markUnread_criteria: MarkUnreadCriteria?
     @Atomic var markUnread_lastReadMessageId: MessageId?
     @Atomic var markUnread_completion: ((Result<ChatChannel, Error>) -> Void)?
     @Atomic var markUnread_completion_result: Result<ChatChannel, Error>?
@@ -170,6 +172,12 @@ final class ChannelUpdater_Mock: ChannelUpdater {
 
         updateChannel_payload = nil
         updateChannel_completion = nil
+        updateChannel_completion_result = nil
+
+        partialChannelUpdate_updates = nil
+        partialChannelUpdate_unsetProperties = nil
+        partialChannelUpdate_completion = nil
+        partialChannelUpdate_completion_result = nil
 
         muteChannel_cid = nil
         muteChannel_expiration = nil
@@ -249,7 +257,7 @@ final class ChannelUpdater_Mock: ChannelUpdater {
         
         markUnread_cid = nil
         markUnread_userId = nil
-        markUnread_messageId = nil
+        markUnread_criteria = nil
         markUnread_lastReadMessageId = nil
         markUnread_completion = nil
         markUnread_completion_result = nil
@@ -330,12 +338,14 @@ final class ChannelUpdater_Mock: ChannelUpdater {
     override func updateChannel(channelPayload: ChannelEditDetailPayload, completion: ((Error?) -> Void)? = nil) {
         updateChannel_payload = channelPayload
         updateChannel_completion = completion
+        updateChannel_completion_result?.invoke(with: completion)
     }
 
     override func partialChannelUpdate(updates: ChannelEditDetailPayload, unsetProperties: [String], completion: ((Error?) -> Void)? = nil) {
         partialChannelUpdate_updates = updates
         partialChannelUpdate_unsetProperties = unsetProperties
         partialChannelUpdate_completion = completion
+        partialChannelUpdate_completion_result?.invoke(with: completion)
     }
 
     override func muteChannel(cid: ChannelId, expiration: Int? = nil, completion: ((Error?) -> Void)? = nil) {
@@ -512,10 +522,10 @@ final class ChannelUpdater_Mock: ChannelUpdater {
         markRead_completion_result?.invoke(with: completion)
     }
 
-    override func markUnread(cid: ChannelId, userId: UserId, from messageId: MessageId, lastReadMessageId: MessageId?, completion: ((Result<ChatChannel, Error>) -> Void)? = nil) {
+    override func markUnread(cid: ChannelId, userId: UserId, from unreadCriteria: MarkUnreadCriteria, lastReadMessageId: MessageId?, completion: ((Result<ChatChannel, any Error>) -> Void)? = nil) {
         markUnread_cid = cid
         markUnread_userId = userId
-        markUnread_messageId = messageId
+        markUnread_criteria = unreadCriteria
         markUnread_lastReadMessageId = lastReadMessageId
         markUnread_completion = completion
         markUnread_completion_result?.invoke(with: completion)