+++ TestTools/StreamChatTestTools/SpyPattern/Spy/CDNClient_Spy.swift
@@ -11,6 +11,9 @@ final class CDNClient_Spy: CDNClient, Spy {
     static var maxAttachmentSize: Int64 { .max }
     var uploadAttachmentProgress: Double?
     var uploadAttachmentResult: Result<URL, Error>?
+    
+    var deleteAttachmentRemoteUrl: URL?
+    var deleteAttachmentResult: Error?
 
     func uploadAttachment(
         _ attachment: AnyChatMessageAttachment,
@@ -43,6 +46,17 @@ final class CDNClient_Spy: CDNClient, Spy {
             DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
                 completion(uploadAttachmentResult.map { UploadedFile(fileURL: $0) })
             }
+        }
+    }
+    
+    func deleteAttachment(
+        remoteUrl: URL,
+        completion: @escaping (Error?) -> Void
+    ) {
+        record()
+        deleteAttachmentRemoteUrl = remoteUrl
+        if let result = deleteAttachmentResult {
+            completion(result)
         }
     }
 }