+++ Tests/StreamChatTests/ChatClient_Tests.swift
@@ -1091,7 +1091,8 @@ private class TestEnvironment {
                     requestEncoder: $1,
                     requestDecoder: $2,
                     attachmentDownloader: $3,
-                    attachmentUploader: $4
+                    attachmentUploader: $4,
+                    cdnClient: $5
                 )
                 return self.apiClient!
             },