+++ CHANGELOG.md
@@ -5,6 +5,42 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🔄 Changed
 
+# [4.94.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.94.0)
+_December 02, 2025_
+
+## StreamChat
+### ✅ Added
+- Add `ChatClient.uploadAttachment(localUrl:progress:)` [#3883](https://github.com/GetStream/stream-chat-swift/pull/3883)
+- Add `ChatClient.deleteAttachment(remoteUrl:)` [#3883](https://github.com/GetStream/stream-chat-swift/pull/3883)
+- Add `CDNClient.deleteAttachment(remoteUrl:)` [#3883](https://github.com/GetStream/stream-chat-swift/pull/3883)
+- Add `heic`, `heif` and `svg` formats to the supported image file types [#3883](https://github.com/GetStream/stream-chat-swift/pull/3883)
+- Add `ChatChannelController.markUnread(from:completion:)` and `Chat.markUnread(from:)` where the from argument is `Date` [#3885](https://github.com/GetStream/stream-chat-swift/pull/3885)
+- Add support for filter tags in channels [#3886](https://github.com/GetStream/stream-chat-swift/pull/3886)
+  - Add `ChatChannel.filterTags`
+  - Add `filterTags` channel list filtering key
+  - Add `filterTags` argument to `ChatChannelController` and `Chat` factory methods in `ChatClient`
+  - Add `filterTags` argument to `ChatChannelController.updateChannel` and `ChatChannelController.partialUpdateChannel`
+  - Add `filterTags` argument to `Chat.update` and `Chat.updatePartial`
+### 🐞 Fixed
+- Fix rare crash in `WebSocketClient.connectEndpoint` [#3882](https://github.com/GetStream/stream-chat-swift/pull/3882)
+- Fix audio recordings not playing from AirPods automatically [#3884](https://github.com/GetStream/stream-chat-swift/pull/3884)
+- Fix audio recordings not using AirPods mic automatically [#3884](https://github.com/GetStream/stream-chat-swift/pull/3884)
+
+## StreamChatUI
+### 🐞 Fixed
+- Fix marking channel read when scrolling to the bottom without unread counts [#3881](https://github.com/GetStream/stream-chat-swift/pull/3881)
+
+# [4.93.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.93.0)
+_November 18, 2025_
+
+## StreamChat
+### ✅ Added
+- Add `ChatClientConfig.isAutomaticSyncOnReconnectEnabled` for toggling automatic syncing [#3879](https://github.com/GetStream/stream-chat-swift/pull/3879)
+### ⚡️ Performance
+- Reduce SDK size by converting Events from structs to classes [#3878](https://github.com/GetStream/stream-chat-swift/pull/3878)
+### 🔄 Changed
+- Change Events from structs to classes [#3878](https://github.com/GetStream/stream-chat-swift/pull/3878)
+
 # [4.92.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.92.0)
 _November 05, 2025_